% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublecrochet.R
\name{combined_collapse_and_save}
\alias{combined_collapse_and_save}
\title{Collapse text and save as .Rmd}
\usage{
combined_collapse_and_save(combined, file = "temp.Rmd")
}
\arguments{
\item{file}{}
}
\value{

}
\description{
Collapse text and save as .Rmd
}
