% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublecrochet.R
\name{crochet}
\alias{crochet}
\title{Title crochet}
\usage{
crochet(input, output = stringr::str_replace(input, "\\\\.rmd|\\\\.Rmd",
  "_double.Rmd"))
}
\arguments{
\item{input}{path to .Rmd file}

\item{output}{path to .Rmd output}
}
\value{
saved file
}
\description{
Title crochet
}
