% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublecrochet.R
\name{data_frame_portioned_collapse_to_portion_list}
\alias{data_frame_portioned_collapse_to_portion_list}
\title{convert to list containing strings that have content between slide breaks}
\usage{
data_frame_portioned_collapse_to_portion_list(data_frame_portioned)
}
\arguments{
\item{data_frame_portioned}{}
}
\value{

}
\description{
convert to list containing strings that have content between slide breaks
}
