% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublecrochet.R
\name{data_frame_portioned_remove_pause_points}
\alias{data_frame_portioned_remove_pause_points}
\title{remove slide breaks from content}
\usage{
data_frame_portioned_remove_pause_points(data_frame_portioned)
}
\arguments{
\item{data_frame_portioned}{}
}
\value{

}
\description{
remove slide breaks from content
}
