% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublecrochet.R
\name{doublecrochet}
\alias{doublecrochet}
\title{Title crochet}
\usage{
doublecrochet(
  input,
  dcrocheted_rmd = stringr::str_replace(input, "\\\\.rmd|\\\\.Rmd",
    "_double_crochet.Rmd"),
  render = T
)
}
\arguments{
\item{input}{path to .Rmd file}

\item{output}{path to .Rmd output}
}
\value{
saved file
}
\description{
Title crochet
}
