% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublecrochet.R
\name{lines_data_frame_portioned}
\alias{lines_data_frame_portioned}
\title{rmd text converted to data frame and slide breaks identified
and which lines associated with each slide}
\usage{
lines_data_frame_portioned(lines, pause_point = "^---$")
}
\arguments{
\item{pause_point}{}
}
\value{

}
\description{
rmd text converted to data frame and slide breaks identified
and which lines associated with each slide
}
