% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublecrochet.R
\name{original_quoted_combine}
\alias{original_quoted_combine}
\title{Alternate original and quoted
quoted first, except for initial}
\usage{
original_quoted_combine(original, quoted)
}
\arguments{
\item{quoted}{}
}
\value{

}
\description{
Alternate original and quoted
quoted first, except for initial
}
