% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublecrochet.R
\name{portion_list_quote}
\alias{portion_list_quote}
\title{Quote the material betweeen the pause points ---}
\usage{
portion_list_quote(portion_list)
}
\arguments{
\item{portion_list}{}
}
\value{

}
\description{
Quote the material betweeen the pause points ---
}
