% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublecrochet.R
\name{quoted_style}
\alias{quoted_style}
\title{Adding style and message to quoted text}
\usage{
quoted_style(quoted)
}
\arguments{
\item{quoted}{}
}
\value{

}
\description{
Adding style and message to quoted text
}
